Extract helpers module
